Date: Wed, 17 Oct 2001 21:26:59 +1000 From: tali streit <astreit@bigpond.net.au>
comparison 0.0.2 - i've included the language targets, the WOS, and the
Software Catalogue that I will soon propose.

For microsoft i have replaced the WSDL/SOAP with Indigo. Was wondering
what they would call it :)

perhaps you can fill me in on what else i should put for DotGNU? I am
not exactly sure where to fit the SEE and the DEE (DEE == WOS, or not?)